chore: move ensure_hex_and_rebar to engine
While the scope of Forge seems a bit under-defined, I don't think it a
good place for a function that only performs side-effects and is only
used by Engine. Thereforce I think Engine.Mix is a better place for it.

Not sure about Forge.Internet, maybe it should be moved as well.
