+++ apps/forge/lib/forge/project.ex
@@ -6,9 +6,6 @@ defmodule Forge.Project do
   as well as business logic for how to change its attributes.
   """
   alias Forge.Document
-  alias Forge.Internet
-
-  require Logger
 
   defstruct root_uri: nil,
             mix_exs_uri: nil,
@@ -391,16 +388,5 @@ defmodule Forge.Project do
 
   def kind(%__MODULE__{} = project) do
     project.kind
-  end
-
-  def ensure_hex_and_rebar do
-    if Internet.connected_to_internet?() do
-      Mix.Task.run("local.hex", ~w(--force --if-missing))
-      Mix.Task.run("local.rebar", ~w(--force --if-missing))
-      :ok
-    else
-      Logger.warning("Could not connect to hex.pm, dependencies will not be fetched")
-      :ok
-    end
   end
 end
